scripts/ml/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
onnx>=1.15.0
onnxscript>=0.1.0
onnxruntime>=1.17.0
numpy>=1.26.0
Pillow>=10.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
